Modified Logic in the Teacher Service Layer

+++ src/main/java/org/example/studentmanagementsystem/service/TeacherService.java
@@ -1,9 +1,13 @@
 package org.example.studentmanagementsystem.service;
 
 import org.example.studentmanagementsystem.entity.Quiz;
+import org.example.studentmanagementsystem.entity.Student;
 import org.example.studentmanagementsystem.entity.Teacher;
 import org.example.studentmanagementsystem.repository.QuizRepository;
 import org.example.studentmanagementsystem.repository.TeacherRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,32 +16,39 @@ import java.util.List;
 public class TeacherService {
     private TeacherRepository teacherRepository;
 
+    @Autowired
     public TeacherService(TeacherRepository teacherRepository){
         this.teacherRepository = teacherRepository;
     }
 
-    public void save(Teacher teacher){
+    public ResponseEntity<Teacher> save(Teacher teacher){
         teacherRepository.save(teacher);
+        return new ResponseEntity<>(teacher, HttpStatus.CREATED);
     }
 
-    public Teacher findById(int id){
+    public ResponseEntity<Teacher> findById(int id){
         Teacher teacher = teacherRepository.findById(id);
-        if(teacher == null){
-            throw new NullPointerException("teacher of id " + id + " doesn't exist");
+        if (teacher == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return teacher;
+        return new ResponseEntity<>(teacher, HttpStatus.OK);
 
     }
 
-    public void delete(int id){
+    public ResponseEntity<Void> delete(int id){
+        if(teacherRepository.findById(id) == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
         teacherRepository.delete(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    public List<Teacher> findAll(){
+    public ResponseEntity<List<Teacher>> findAll(){
         List<Teacher> teachers = teacherRepository.findAll();
-        if(teachers == null)
-            throw new NullPointerException("List of teachers is empty and currently NULL");
-        return teachers;
+        if (teachers.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(teachers, HttpStatus.OK);
     }
 
     public void update(Teacher teacher){
